+++ internal/handlers/deleteHandler.go
