+++ internal/handlers/postHandler_test.go
@@ -2,10 +2,12 @@ package handlers
 
 import (
 	"io"
+	"log"
 	"net/http"
 	"net/http/httptest"
 	"strings"
 	"testing"
+	"time"
 
 	"github.com/scaranin/go-svc-short-url/internal/auth"
 	"github.com/scaranin/go-svc-short-url/internal/config"
@@ -13,73 +15,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
-
-func TestURLHandler_GetHandle(t *testing.T) {
-	type want struct {
-		statusCode  int
-		request     string
-		location    string
-		contentType string
-	}
-	tests := []struct {
-		name string
-		want want
-	}{
-		{
-			name: "get handle negative test #1",
-			want: want{
-				statusCode:  http.StatusBadRequest,
-				request:     "http://localhost:8080",
-				location:    "",
-				contentType: "text/plain",
-			},
-		},
-
-		{
-			name: "get handle positive test #1",
-			want: want{
-				statusCode:  http.StatusTemporaryRedirect,
-				request:     "http://localhost:8080/pkmdI_i-nYcS6P7hSfjTtWUmfcA=",
-				location:    "https://practicum.yandex.ru/",
-				contentType: "text/plain",
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			cfg, err := config.CreateConfig()
-			if err != nil {
-				return
-			}
-			store, err := storage.CreateStoreFile(cfg.FileStoragePath)
-			if err != nil {
-				return
-			}
-			defer store.Close()
-			h1 := CreateHandle(cfg, store, auth.NewAuthConfig())
-
-			reqPost := httptest.NewRequest(http.MethodPost, "/", strings.NewReader(tt.want.location))
-			reqPost.Header.Set("Content-Type", tt.want.contentType)
-			recPost := httptest.NewRecorder()
-			h1.PostHandle(recPost, reqPost)
-
-			reader := strings.NewReader(``)
-			client := &http.Client{}
-			req := httptest.NewRequest(http.MethodGet, tt.want.request, reader)
-
-			req.Header.Add("Content-Type", tt.want.contentType)
-
-			res, err := client.Do(req)
-			if err != nil {
-				return
-			}
-			defer res.Body.Close()
-			assert.Equal(t, tt.want.statusCode, res.StatusCode)
-			assert.Equal(t, tt.want.location, res.Header.Get("Location"))
-		})
-	}
-
-}
 
 func TestURLHandler_PostHandle(t *testing.T) {
 	type want struct {
@@ -202,85 +137,41 @@ func TestURLHandler_PostHandleJson(t *testing.T) {
 	}
 }
 
-func TestURLHandler_PingHandle(t *testing.T) {
-	type want struct {
-		statusCode  int
-		request     string
-		location    string
-		contentType string
-	}
-	tests := []struct {
-		name string
-		want want
-	}{
-		{
-			name: "ping handle positive test #1",
-			want: want{
-				statusCode:  http.StatusOK,
-				request:     "http://localhost:8080/ping",
-				location:    "",
-				contentType: "text/plain",
-			},
-		},
+func BenchmarkPost(b *testing.B) {
+	cfg, err := config.CreateConfig()
+	if err != nil {
+		log.Fatal(err)
 	}
 
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			reader := strings.NewReader(``)
-			client := &http.Client{}
-			req := httptest.NewRequest(http.MethodGet, tt.want.request, reader)
-
-			req.Header.Add("Content-Type", tt.want.contentType)
-
-			res, err := client.Do(req)
-			if err != nil {
-				return
-			}
-			defer res.Body.Close()
-
-			assert.Equal(t, tt.want.statusCode, res.StatusCode)
-
-		})
+	store, err := config.CreateStore(cfg)
+	if err != nil {
+		log.Fatal(err)
 	}
-}
 
-func TestURLHandler_DeleteHandle(t *testing.T) {
-	type want struct {
-		statusCode  int
-		request     string
-		location    string
-		contentType string
+	auth := auth.NewAuthConfig()
+
+	h := CreateHandle(cfg, store, auth)
+
+	req := httptest.NewRequest(http.MethodPost, h.BaseURL, nil)
+
+	w := httptest.NewRecorder()
+
+	resAuthToken, err := auth.BuildJWTString()
+	if err != nil {
+		log.Fatal(err)
 	}
-	tests := []struct {
-		name string
-		want want
-	}{
-		{
-			name: "delete handle positive test #1",
-			want: want{
-				statusCode:  http.StatusBadRequest,
-				request:     "http://localhost:8080/api/user/urls",
-				location:    "",
-				contentType: "text/plain",
-			},
-		},
+
+	cookie := &http.Cookie{
+		Name:     auth.CookieName,
+		Value:    resAuthToken,
+		Expires:  time.Now().Add(auth.TokenExp),
+		HttpOnly: true,
+		Path:     "/",
 	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			reader := strings.NewReader(``)
-			client := &http.Client{}
-			req := httptest.NewRequest(http.MethodDelete, tt.want.request, reader)
+	req.AddCookie(cookie)
 
-			req.Header.Add("Content-Type", tt.want.contentType)
-
-			res, err := client.Do(req)
-			if err != nil {
-				return
-			}
-			defer res.Body.Close()
-
-			assert.Equal(t, tt.want.statusCode, res.StatusCode)
-
-		})
+	for i := 0; i < b.N; i++ {
+		h.post(w, req, "application/json")
 	}
+
 }
